feat: partial migration to tailwind and removal of htmx (unfinished)
