+++ app.js
@@ -1,13 +1,14 @@
 import express from 'express';
 import path from 'path';
 import bodyParser from 'body-parser';
-import SteamAuth from 'node-steam-openid';
 import session from 'express-session';
 import sqlite3 from 'sqlite3';
 import { renderFile } from 'ejs';
 import { fileURLToPath } from 'url';
-import moment from 'moment';
-import {appendRequestToFile, logRequestToFile} from './helpers/logger.js'
+import forumPostRoutes from './routes/forumPosts.js';
+import steamRoutes from './routes/steamAuth.js';
+import signupRoutes from './routes/signup.js';
+import db from './db.js';
 
 const __filename = fileURLToPath(import.meta.url);
 const __dirname = path.dirname(__filename);
@@ -28,332 +29,10 @@ app.use(session({
     saveUninitialized: true
 }));
 
-app.use((req, res, next) => {
-    const noRedirectUrls = ['openid', 'player_page', '2v2cup', 'discord'];
-    const shouldRedirect = !noRedirectUrls.some(substring => req.url.includes(substring)) 
-                            && req.headers['hx-request'] !== 'true';
-    if (req.url !== '/' && shouldRedirect) {
-        return res.redirect('/');
-    }
-    next();
-});
-// user database
-const db = new sqlite3.Database('users.db', sqlite3.OPEN_READWRITE, async (err) => {
-    if (err) {
-        console.error(err.message);
-    }
-    await db.run(`CREATE TABLE IF NOT EXISTS users (
-            id INTEGER PRIMARY KEY AUTOINCREMENT,
-            steam_id TEXT NOT NULL,
-            steam_username TEXT NOT NULL)`);
 
-    await db.run(`CREATE TABLE IF NOT EXISTS threads (
-                  id INTEGER PRIMARY KEY AUTOINCREMENT,
-                  title TEXT NOT NULL,
-                  content TEXT NOT NULL,
-                  created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
-                  bumped_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
-
-                  owner INTEGER NOT NULL,
-                  FOREIGN KEY (owner) REFERENCES users(id))`);
-
-    await db.run(`CREATE TABLE IF NOT EXISTS posts (
-                  id INTEGER PRIMARY KEY AUTOINCREMENT,
-                  content TEXT NOT NULL,
-                  created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
-
-                  thread INTEGER NOT NULL,
-                  owner INTEGER NOT NULL,
-                  FOREIGN KEY (owner) REFERENCES users(id)
-                  FOREIGN KEY (thread) REFERENCES thread(id))`);
-
-    await db.run(`CREATE TABLE IF NOT EXISTS activity (
-                id INTEGER PRIMARY KEY AUTOINCREMENT,
-                thread_count INTEGER NOT NULL,
-                post_count INTEGER NOT NULL,
-                period INTEGER NOT NULL,
-
-                owner INTEGER NOT NULL,
-                FOREIGN KEY (owner) REFERENCES users(id))`);
-
-    // await db.run(`CREATE TABLE IF NOT EXISTS articles (
-    //             id INTEGER PRIMARY KEY AUTOINCREMENT,
-    //             title TEXT NOT NULL,
-    //             content TEXT NOT NULL,
-    //             owner INTEGER NOT NULL,
-    //             created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
-    //             FOREIGN KEY (owner) REFERENCES users(id))`);
-
-    await db.run(`CREATE TABLE IF NOT EXISTS moderators (
-                id INTEGER PRIMARY KEY AUTOINCREMENT,
-                user_id INTEGER NOT NULL,
-                FOREIGN KEY (user_id) REFERENCES users(id))`);
-
-    const moderatorIds = ['76561198082657536', '76561198041183975'];
-    db.all(`SELECT users.*, moderators.id as moderator_id
-                    FROM moderators
-                    JOIN users ON moderators.user_id = users.id`, [], (err, moderators) => {
-        if (err) {
-            console.error('Error querying database: ' + err.message);
-        }
-        const match = moderators.some(moderator => moderatorIds.includes(moderator.steam_id));
-
-        if (!match) {
-            db.run(`INSERT INTO moderators (user_id)
-            SELECT id FROM users WHERE steam_id IN ('76561198082657536', '76561198041183975')`);
-        }
-    });
-});
-
-// 3 posts per hour, 30 replies
-const MAX_POSTS = 3;
-
-// 1 hour
-const REFRESH_PERIOD = 3600000;
-
-app.post('/postContent', async (req, res) => {
-    const { content, title } = req.body;
-    const user = req.session.user;
-
-    if (!user) {
-        return res.status(401).json({ error: 'Login to create post' });
-    }
-
-    if (!content || !title) {
-        return res.status(400).json({ error: 'Content and title are required' });
-    }
-
-    console.log(user);
-    db.get('SELECT id FROM users WHERE steam_id = ?', [user.steamid], (err, row) => {
-        if (err) {
-            console.error(err.message);
-            return res.status(500).json({ error: 'Failed to post content.' });
-        }
-
-        const user_id = row.id;
-
-        db.get('SELECT * FROM activity WHERE owner = ?', [user_id], (err, activityRow) => {
-            if (err) {
-                console.error(err.message);
-                return res.status(500).json({ error: 'Failed to retrieve activity data.' });
-            }
-
-            const current_time = Date.now();
-            let threadCount = 0;
-            let lastPeriod = current_time;
-
-            if (activityRow) {
-                threadCount = activityRow.thread_count;
-                lastPeriod = activityRow.period;
-
-                if (current_time - lastPeriod >= REFRESH_PERIOD) {
-                    threadCount = 0;
-                    lastPeriod = current_time;
-                }
-            }
-
-            if (threadCount >= MAX_POSTS) {
-                console.log('You have reached max numebr of posts')
-                return res.status(400).json({ error: 'You have reached the limit of posts. Try again later. ' })
-            }
-
-            db.run('INSERT INTO threads (title, content, owner) VALUES (?, ?, ?)', [title, content, user_id], (err) => {
-                if (err) {
-                    console.error(err.message);
-                    return res.status(500).json({ error: 'Failed to post content.' });
-                } else {
-                    console.log(`Forum Post "${content}" saved to the database.`);
-
-                    if (activityRow) {
-                        db.run('UPDATE activity SET thread_count = ?, period = ? WHERE owner = ?', [threadCount + 1, lastPeriod, user_id])
-                    } else {
-                        db.run('INSERT INTO activity (thread_count, post_count, period, owner) VALUES (?, ?, ?, ?)', [1, 0, current_time, user_id]);
-                    }
-
-                    return res.status(200).redirect('/forums');
-                }
-            });
-        });
-    });
-});
-
-app.get('/forums', (req, res) => {
-    const session = req.session;
-
-    const sql = `SELECT threads.id as thread_id, threads.*, users.steam_username, users.steam_id, users.steam_avatar 
-                 FROM threads 
-                 LEFT JOIN users ON threads.owner = users.id 
-                 ORDER BY bumped_at DESC`;
-
-    db.all(sql, [], (err, forumRows) => {
-        if (err) {
-            console.error(err.message);
-            return res.status(500).json({ error: 'Failed to fetch posts.' });
-        }
-
-        const forums = forumRows.map(forum => {
-            try {
-                const createdAt = moment(forum.created_at);
-                forum.formatted_date = createdAt.format('MM/DD/YYYY HH:mm:ss');
-            } catch (error) {
-                console.error('Error formatting date:', error);
-            }
-            return forum;
-        });
-
-        db.all(`SELECT users.*, moderators.id as moderator_id
-                FROM moderators
-                JOIN users ON moderators.user_id = users.id`, [], (err, moderators) => {
-            if (err) {
-                console.error('Error querying database: ' + err.message);
-                return res.status(500).send('Internal Server Error');
-            }
-
-            res.render('forums', { forums: forums, session: session, moderators: moderators });
-        });
-    });
-});
-
-
-app.post('/posts', (req, res) => {
-    const { content, thread_id } = req.body;
-    const user = req.session.user;
-
-    console.log(req.body);
-
-    if (!user) {
-        return res.status(401).json({ error: 'Login to create post' });
-    }
-
-    if (!content) {
-        return res.status(400).json({ error: 'Content is required' });
-    }
-
-    db.get('SELECT id FROM users WHERE steam_id = ?', [user.steamid], (err, row) => {
-        if (err) {
-            console.error(err.message);
-            return res.status(500).json({ error: 'Failed to post content.' });
-        }
-        const user_id = row.id;
-
-        db.get('SELECT * FROM activity WHERE owner = ?', [user_id], (err, activityRow) => {
-            if (err) {
-                console.error(err.message);
-                return res.status(500).json({ error: 'Failed to retrieve activity data.' });
-            }
-
-            const current_time = Date.now();
-            let postCount = 0;
-            let lastPeriod = current_time;
-
-            if (activityRow) {
-                postCount = activityRow.post_count;
-                lastPeriod = activityRow.period;
-
-                if (current_time - lastPeriod >= REFRESH_PERIOD) {
-                    postCount = 0;
-                    lastPeriod = current_time;
-                }
-            }
-
-            if (postCount >= MAX_POSTS) {
-                console.log('You have reached max numebr of posts')
-                return res.status(400).json({ error: 'You have reached the limit of posts. Try again later. ' })
-            }
-
-            db.run('INSERT INTO posts (content, thread, owner) VALUES (?, ?, ?)', [content, thread_id, user_id], (err) => {
-                if (err) {
-                    console.error(err.message);
-                    return res.status(500).json({ error: 'Failed to post content.' });
-                } else {
-                    console.log(`Post "${content}" saved to the database.`);
-
-                    if (activityRow) {
-                        db.run('UPDATE activity SET post_count = ?, period = ? WHERE owner = ?', [postCount + 1, lastPeriod, user_id])
-                    } else {
-                        db.run('INSERT INTO activity (thread_count, post_count, period, owner) VALUES (?, ?, ?, ?)', [1, 0, current_time, user_id]);
-                    }
-
-                    return res.status(200).redirect('/post/' + thread_id);
-                }
-            });
-        });
-    });
-});
-
-app.post('/remove_post', (req, res) => {
-    ;
-    const postId = req.body.post_id;
-    const mod = req.body.is_moderator;
-    const owner = req.body.is_owner;
-
-    console.log(owner);
-    console.log(mod);
-    if (!owner || !mod) {
-        return res.status(403).send('Unauthorized');
-    }
-
-    const sqlCheckPost = 'SELECT * FROM threads WHERE id = ?';
-    db.get(sqlCheckPost, [postId], (err, post) => {
-        if (err) {
-            console.error(err.message);
-            return res.status(500).send('Failed to verify post.');
-        }
-
-        if (!post) {
-            return res.status(404).send('Post not found.');
-        }
-
-        const sqlDeletePost = 'DELETE FROM threads WHERE id = ?';
-        db.run(sqlDeletePost, [postId], (err) => {
-            if (err) {
-                console.error(err.message);
-                return res.status(500).send('Failed to delete post.');
-            }
-            return res.redirect('/');
-        });
-    });
-});
-
-app.post('/remove_reply', (req, res) => {
-    ;
-    const replyId = req.body.reply_id;
-    const mod = req.body.is_moderator;
-    const owner = req.body.is_owner;
-
-    console.log(owner);
-    console.log(mod);
-    console.log(replyId);
-
-    if (!owner || !mod) {
-        return res.status(403).send('Unauthorized');
-    }
-
-    const sqlCheckPost = 'SELECT * FROM posts WHERE id = ?';
-    db.get(sqlCheckPost, [replyId], (err, post) => {
-        if (err) {
-            console.error(err.message);
-            return res.status(500).send('Failed to verify post.');
-        }
-
-        if (!post) {
-            return res.status(404).send('Post not found.');
-        }
-
-        const sqlDeletePost = 'DELETE FROM posts WHERE id = ?';
-        console.log('3')
-        db.run(sqlDeletePost, [replyId], (err) => {
-            console.log('4')
-            if (err) {
-                console.log('5')
-                console.error(err.message);
-                return res.status(500).send('Failed to delete post.');
-            }
-            console.log('6')
-            return res.redirect('/post/' + post.thread);
-        });
-    });
-});
+app.use('/', forumPostRoutes);
+app.use('/', steamRoutes);
+app.use('/signup', signupRoutes);
 
 const eloDb = new sqlite3.Database('sourcemod-local.sq3', (err) => {
     if (err) {
@@ -367,49 +46,15 @@ app.get('/', (req, res) => {
             console.error('Error querying database: ' + err.message);
             res.status(500).send('Internal Server Error');
         } else {
-            res.render('index', { session: req.session, elo: rows });
+            res.render('layout', {body: 'index', title: 'home', session: req.session, elo: rows });
         }
     });
 });
 
-const API_KEY = '2C7E4CDF46C4D4FB5875A8E6E040BFC0';
-const domain = process.env.DOMAIN || 'http://localhost:3005/';
+app.get('/signup', (req, res) => {
+    res.render('layout', {title: 'Signup', body: 'signup', session: req.session});
+})
 
-const steam = new SteamAuth({
-    realm: domain,
-    returnUrl: domain + 'verify',
-    apiKey: API_KEY,
-});
-
-
-app.get('/init-openid', async (req, res) => {
-    const redirectUrl = await steam.getRedirectUrl();
-    return res.redirect(redirectUrl);
-});
-
-app.get('/verify', async (req, res) => {
-    try {
-        const user = await steam.authenticate(req);
-        req.session.user = user;
-
-        // check if first time login
-        db.get('SELECT * FROM users WHERE steam_id = ?', [user.steamid], (err, row) => {
-            if (err) {
-                console.error('Error querying database: ' + err.message);
-            } else if (!row) {
-                // add if first time
-                db.run('INSERT INTO users (steam_id, steam_username, steam_avatar) VALUES (?, ?, ?)', [user.steamid, user.username, user.avatar.small], (err) => {
-                    if (err) {
-                        console.error('Error inserting into database: ' + err.message);
-                    }
-                });
-            }
-        });
-    } catch (err) {
-        console.error('Authentication error: ' + err.message);
-    }
-    return res.redirect('/');
-});
 
 app.get('/player_page/:steamid', (req, res) => {
     const steamid = req.params.steamid;
@@ -421,41 +66,8 @@ app.get('/player_page/:steamid', (req, res) => {
             const name = req.query.name;
             res.render('empty_player_page', { steamid, name });
         } else {
-            res.render('player_page', { user: row });
+           res.render('layout', {body: 'player_page', title: row.steam_username, user: row, session: req.session });
         }
-    });
-});
-
-app.get('/post/:forumid', (req, res) => {
-    const session = req.session;
-    const forumid = req.params.forumid;
-
-    db.get('SELECT threads.id as thread_id, * FROM threads LEFT JOIN users ON threads.owner = users.id WHERE threads.id = ?', [forumid], (err, row) => {
-        if (err) {
-            console.error('Error querying database: ' + err.message);
-            return res.status(500).send('Internal Server Error');
-        }
-
-        db.all(`SELECT posts.*, users.steam_username, users.steam_id, users.steam_avatar
-            FROM posts LEFT JOIN users ON 
-            posts.owner = users.id WHERE thread = ?`, [forumid], (err, posts) => {
-
-            if (err) {
-                console.error('Error querying database: ' + err.message);
-                return res.status(500).send('Internal Server Error');
-            }
-
-            db.all(`SELECT users.*, moderators.id as moderator_id
-                    FROM moderators
-                    JOIN users ON moderators.user_id = users.id`, [], (err, moderators) => {
-                if (err) {
-                    console.error('Error querying database: ' + err.message);
-                    return res.status(500).send('Internal Server Error');
-                }
-
-                res.render('posts', { session: session, moderators: moderators, posts: posts, thread: row });
-            });
-        });
     });
 });
 
@@ -476,29 +88,12 @@ app.get('/users', (req, res) => {
                 return;
             }
 
-            res.render('users', { users: users, moderators: moderators });
+            res.render('layout', { title: 'Users', body: 'users', users: users, moderators: moderators, session: req.session });
         });
     });
 });
 
 app.get('/click_user_steamid')
-
-app.get('/load', (req, res) => {
-    const { page } = req.query;
-    if (page === 'main') {
-        eloDb.all('SELECT * FROM mgemod_stats ORDER BY rating DESC', [], (err, rows) => {
-            if (err) {
-                console.error('Error querying database: ' + err.message);
-                res.status(500).send('Internal Server Error');
-            } else {
-                return res.render('main', { elo: rows });
-            }
-        });
-    } else {
-        res.render(page);
-    }
-});
-
 
 app.get('/2v2cup', (req, res) => {
     res.render('2v2cup');
@@ -508,64 +103,6 @@ app.get('/discord', (req, res) => {
     const discordInviteLink = 'https://discord.gg/j6kDYSpYbs'; 
     res.redirect(discordInviteLink);
 });
-
-// app.get('/create_article', (req, res) => {
-//     if (!req.session.user) {
-//         return res.redirect('/login');
-//     }
-//     res.render('create_article');
-// });
-
-// app.post('/create_article', (req, res) => {
-//     const { title, content } = req.body;
-//     const user = req.session.user;
-
-//     if (!user) {
-//         return res.status(401).json({ error: 'Login to create article' });
-//     }
-
-//     if (!title || !content) {
-//         return res.status(400).json({ error: 'Title and content are required' });
-//     }
-
-//     db.run('INSERT INTO articles (title, content, owner) VALUES (?, ?, ?)', [title, content, user.id], (err) => {
-//         if (err) {
-//             console.error(err.message);
-//             return res.status(500).json({ error: 'Failed to create article.' });
-//         }
-//         return res.redirect('/articles'); // Redirect to articles page after creation
-//     });
-// });
-
-// app.get('/articles', (req, res) => {
-//     db.all(`SELECT articles.*, users.steam_username AS author 
-//             FROM articles 
-//             LEFT JOIN users ON articles.owner = users.id 
-//             ORDER BY articles.created_at DESC`, [], (err, rows) => {
-//         if (err) {
-//             console.error(err.message);
-//             return res.status(500).json({ error: 'Failed to fetch articles.' });
-//         }
-//         res.render('articles', { articles: rows });
-//     });
-// });
-
-// app.get('/article/:id', (req, res) => {
-//     const articleId = req.params.id;
-//     db.get(`SELECT articles.*, users.steam_username AS author 
-//             FROM articles 
-//             LEFT JOIN users ON articles.owner = users.id 
-//             WHERE articles.id = ?`, [articleId], (err, row) => {
-//         if (err) {
-//             console.error(err.message);
-//             return res.status(500).json({ error: 'Failed to fetch article.' });
-//         }
-//         if (!row) {
-//             return res.status(404).json({ error: 'Article not found.' });
-//         }
-//         res.render('article', { article: row });
-//     });
-// });
 
 app.get('/logout', (req, res) => {
     req.session.destroy(err => {
