+++ package.json
@@ -25,6 +25,7 @@
   "devDependencies": {
     "@eslint/js": "^9.5.0",
     "eslint": "9.x",
-    "globals": "^15.5.0"
+    "globals": "^15.5.0",
+    "tailwindcss": "^3.4.4"
   }
 }
