+++ tailwind.config.js
@@ -1,0 +1,9 @@
+/** @type {import('tailwindcss').Config} */
+export default {
+  content: ['./views/**/*.{ejs,html}'],
+  theme: {
+    extend: {},
+  },
+  plugins: [],
+}
+
