+++ views/css/tw.css
@@ -1,0 +1,24 @@
+@tailwind base;
+@tailwind components;
+@tailwind utilities;
+
+.bg-custom-background {
+  background-image: url('/images/badlands-mid.png');
+  background-size: cover;
+  background-position: center;
+}
+
+
+.sidebar {
+  position: fixed;
+  top: 0;
+  left: -300px;
+  width: 300px;
+  height: 100vh;
+  background-color: #1a202c;
+  transition: left 0.3s ease;
+}
+
+.sidebar.open {
+  left: 0;
+}
