+++ views/main.ejs
@@ -1,7 +1,8 @@
 <div class="main-overlay">
     <h1 class="mge-title">
         <img src="images/mge_transparent_logo.png" class="mge-logo">
-         REWRITING MGE</h1>
+        REWRITING MGE
+    </h1>
     <div class="placeholder">
         <div class="elo-search-container">
             <input type="text" id="eloSearch" placeholder="Search by username...">
@@ -45,8 +46,10 @@
     <div class="main-menu">
         <div class="main-menu-buttons">
             <a class="main-menu-button main-1" hx-get="/users" hx-target="#main" hx-swap="innerHtml">Users</a>
-            <a class="main-menu-button main-3" hx-get="/load?page=league" hx-target="#main" hx-swap="innerHtml">League</a>
-            <a class="main-menu-button main-2" hx-get="/load?page=seasons" hx-target="#main" hx-swap="innerHtml">Seasons</a>
+            <a class="main-menu-button main-3" hx-get="/load?page=league" hx-target="#main"
+                hx-swap="innerHtml">League</a>
+            <a class="main-menu-button main-2" hx-get="/load?page=seasons" hx-target="#main"
+                hx-swap="innerHtml">Seasons</a>
             <a class="main-menu-button main-4" hx-get="/forums" hx-swap="innerHtml" hx-target="#main">Forums</a>
         </div>
