+++ views/player_page.ejs
@@ -1,30 +1,20 @@
-<div class="dashboard-container">
-    <link rel="stylesheet" href="css/player-page.css" />
-    <title>User Profile</title>
-
-    <h1>
-        <%= user.steam_username %>
-    </h1>
-    <img src='<%= user.steam_avatar %>' alt="Avatar" class="player-page-profile-pic">
-
-    <div class="logos-container">
-        <div class="logo">
-            <a href="https://steamcommunity.com/profiles/<%= user.steam_id %>" target="_blank">
-                <img src="https://upload.wikimedia.org/wikipedia/commons/8/87/New_Steam_Logo_with_name.jpg"
-                    alt="Steam Logo" class="steam-profile-logo">
-            </a>
-        </div>
-        <div class="logo">
-            <a href="https://logs.tf/profile/<%= user.steam_id %>" target="_blank">
-                <img src="/images/logs-tf-logo.png" alt="Logs.tf Logo" class="logs-profile-logo">
-            </a>
-        </div>
-        <div class="logo rgl-logo">
-            <a href="https://rgl.gg/Public/PlayerProfile?p=<%= user.steam_id %>&r=40" target="_blank">
-                <img src="/images/rgl-logo.png" alt="RGL.gg Logo" class="rgl-profile-logo">
-            </a>
-        </div>
+<div class="text-black flex flex-col mx-auto max-w-5xl w-3/5 bg-white shadow-md rounded-lg p-6">
+    <div class="flex flex-col justify-center align-middle items-center">
+        <h1>
+            <%= user.steam_username %>
+        </h1>
+        <img class="h-40 rounded-md object-contain" src='<%= user.steam_avatar %>' alt="Avatar"
+            class="player-page-profile-pic">
     </div>
 
-    <a href="/logout">Logout</a>
+    <a href="https://steamcommunity.com/profiles/<%= user.steam_id %>" target="_blank">
+        <img class="h-20" src="https://upload.wikimedia.org/wikipedia/commons/8/87/New_Steam_Logo_with_name.jpg"
+            alt="Steam Logo" class="steam-profile-logo">
+    </a>
+    <a href="https://logs.tf/profile/<%= user.steam_id %>" target="_blank">
+        <img class="h-20" src="/images/logs-tf-logo.png" alt="Logs.tf Logo" class="logs-profile-logo">
+    </a>
+    <a href="https://rgl.gg/Public/PlayerProfile?p=<%= user.steam_id %>&r=40" target="_blank">
+        <img class="h-20" src="/images/rgl-logo.png" alt="RGL.gg Logo" class="rgl-profile-logo">
+    </a>
 </div>
