+++ views/sidebar.ejs
@@ -1,0 +1,36 @@
+<div id="sidebar" class="sidebar">
+  <!-- Sidebar content here -->
+  <header class="overlay">
+    <div class="top-section">
+      <h1 class="title">
+        <a href="/">mge.tf</a>
+      </h1>
+      <img class="logo" src="">
+    </div>
+    <div class="bottom-section">
+      <div hx-boost="true" class="menu">
+        <a href="/">Home</a>
+        <a href="/users">Users</a>
+        <div class="dropdown">
+          <a href="/league">League</a>
+          <div class="dropdown-content">
+            <a href="/league">League Home</a>
+            <a href="/signup">Signups</a>
+            <a href="/rules">Rules</a>
+          </div>
+        </div>
+        <div class="dropdown">
+          <a class="menu-button dropdown" href="/load?page=seasons" hx-target="#main" hx-swap="innerHtml">Seasons ▶</a>
+          <div class="dropdown-content">
+            <a class="dropdown-option" href="/load?page=season1" hx-target="#main" hx-swap="innerHtml">Season 1</a>
+          </div>
+        </div>
+        <a href="/forums">Forums</a>
+        <a href="/2v2cup">2v2 Cup</a>
+      </div>
+    </div>
+  </header>
+</div>
+
+<script src="/js/main.js"></script>
+<script src="/js/sidebar.js"></script>
