+++ views/signup.ejs
@@ -1,9 +1,6 @@
 <h1 class="description">CHOOSE ONE</h1>
-<div class="background-page"></div>
 
-<link rel="stylesheet" href="css/signup.css" />
-
-<div hx-boost="true" class="signup-button-container">
-    <a class="custom-button button-blue" href="/load?page=1v1signup.ejs" hx-swap="innerHtml" hx-target="#main">1v1</a>
-    <a class="custom-button button-red" href="/load?page=2v2signup.ejs" hx-swap="innerHtml" hx-target="#main">2v2</a>
+<div class="signup-button-container">
+    <a class="custom-button button-blue" href="/signup/1v1signup">1v1</a>
+    <a class="custom-button button-red" href="/signup/2v2signup">2v2</a>
 </div>W
