+++ views/users.ejs
@@ -13,7 +13,7 @@
                 displayName = displayName.slice(0, maxLength) + '...';
                 }
                 %>
-                <a class="users_username" href="#">
+                <a class="users_username" href="/player_page/<%= user.steam_id %>">
                     <%= displayName %>
                 </a>
                 <span class="users_steam_id">
